Address review: rename k8s_config -> k8s_kubeconfig, drop unused cli helpers
Per Alberto's review on #290:
- Rename k8s_config.py -> k8s_kubeconfig.py ("config" was ambiguous; the
  module is about the kubeconfig). Updated the 4 import sites.
- Remove unused cli_utils helpers gen_argparse and assert_equals (no
  callers; gen_argparse was a NotImplementedError stub). get_flag_value
  stays — it's used by the waku regression experiment.
