+++ src/utils/cli_utils.py
@@ -1,16 +1,6 @@
 # Python Imports
-import argparse
 import re
 from typing import List, Optional
-
-
-def gen_argparse(arg_defs):
-    parser = argparse.ArgumentParser(description="Args generated from template.")
-    # TODO [mutually exclusive args]: add logic here
-    for arg in arg_defs:
-        kwargs = {key: value for key, value in arg.items() if key != "name"}
-        parser.add_argument(arg["name"], **kwargs)
-    raise NotImplementedError()
 
 
 def get_flag_value(flag: str, command: List[str]) -> Optional[int]:
@@ -21,7 +11,3 @@ def get_flag_value(flag: str, command: List[str]) -> Optional[int]:
         except (TypeError, IndexError):
             pass
     return None
-
-
-def assert_equals(obj_1, obj_2):
-    assert obj_1 == obj_2, f"Assertion failed: `{obj_1}` == `{obj_2}`"
